Scripts/Gilneas: added quest fixes for 14293, and 14154.
Scripts/Gilneas: coding standards and clen-up done.

this clean-up and change of the way it was coded was to shut down any QQ that we've had in past commits...

sry it was committed this way... but i wanted to show that what was said to these select few in comments "WAS NOT" an open/empty threat, on my part.


Former-commit-id: d3ee56f59c3704bce35ad8d680ab5e580b277d98